refactor(appbar): move select instance themetoggle profile to upper bar
